Change revenue from int to long.

+++ app/src/main/java/com/github/pedramrn/slick/parent/datasource/network/models/BoxOfficeItem.java
@@ -14,7 +14,7 @@ import com.google.gson.annotations.SerializedName;
 @AutoValue
 public abstract class BoxOfficeItem {
     @SerializedName("revenue")
-    public abstract Integer revenue();
+    public abstract Long revenue();
 
     @SerializedName("movie")
     public abstract MovieTraktMetadata movie();
